Remove generic parsing in favour of specific ODS parsing.
The ODS XML format is odd in many respects. Firstly, values are often recorded as a value attribute, rather than the content of a tag. Secondly, the model isn’t logical, so that one can have multiple dates but each is a top-level element, but other to-many relationships are explicitly held within a to-many tag such as <succs> for successor organisations. This means writing a generically useful parser is more difficult; instead, this specifically parses the XML and so provides guarantees as to the *output* of the parsing - including consistency - which was hard to get right if one organisation provided multiple dates and another one date. These should be either separate data items (legalDate and operationalDate - this is what is implemented now) or an explicit model of dates containing one or more date items. I’ve adopted the former as there are only two current “date types” and if we’re going to make use of them computationally, let’s make their type explicit rather than an attribute.