Add resolver for 'isPartOf'
The determination of a hierarchy of organisations isn't very easy. There are a number of relationship types, and their use depends on the role of the source organisation. The authoritative source of information on this is an ancillary spreadsheet available from https://digital.nhs.uk/services/organisation-data-service/technical-guides/guidance-for-developers/xml-organisation-data-products-support-products-and-tools#xml-support-products  which provides a table of rules for each role. However, we can simplify this by looking for certain types of relationship and using the first relationship that can be taken to mean "isPartOf".